backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.3
django-redis==5.4.0
django-celery-beat==2.6.0
django-channels[daphne]==4.2.0
channels-redis==4.2.0
django-parler==2.3
django-encrypted-model-fields==0.6.5
django-auditlog==3.0.0
django-prometheus==2.3.1
django-storages==1.14.2
django-import-export==4.0.7
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
gunicorn==21.1.0
daphne==4.1.0
Pillow==10.2.0
pandas==2.1.5
numpy==1.26.3
scikit-learn==1.4.0
stripe==7.12.0
python-dotenv==1.0.1
whitenoise==6.6.0
boto3==1.34.14
pycryptodome==3.20.0
qrcode==7.4.2
python-barcode==0.15.1
openpyxl==3.1.2
reportlab==4.0.9
weasyprint==61.2
pytest==7.4.4
pytest-django==4.7.0
pytest-asyncio==0.23.3
factory-boy==3.3.1
faker==22.2.0
coverage==7.4.0
